requirements.txt
grpcio~=1.57.0
grpcio-tools~=1.57.0
pillow~=11.0.0
Flask~=3.1.0
torch~=2.5.1
opencv-python~=4.9.0.80
numpy~=1.26.4
psutil~=6.1.0
mysql~=0.0.3
mysql-connector-python~=9.1.0
requests~=2.32.3
urllib3~=2.2.3
tabulate~=0.9.0
aiohttp~=3.11.11
torchvision~=0.20.1
tqdm~=4.67.1